real_demo/collision_free_ik/pyproject.toml
[build-system]
requires = ["flit_core >=2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "mink"
version = "0.0.11"
readme = "README.md"
authors = [
    {name = "Kevin Zakka", email = "zakka@berkeley.edu"},
]
requires-python = ">=3.8"
dynamic = ["description"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Robot Framework :: Library",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    # 3.1.6 for mj_geomDistance.
    "mujoco >= 3.1.6",
    "qpsolvers[daqp] >= 4.3.1",
    "typing_extensions",
    "numpy",
]
keywords = ["inverse", "kinematics", "mujoco"]

[project.optional-dependencies]
examples = [
    "loop-rate-limiters >= 0.1.0",
    "qpsolvers[daqp] >= 4.3.1",
    "osqp >=0.6.2,<1",
]
test = [
    "absl-py",
    "coveralls",
    "pytest",
    "robot_descriptions >= 1.9.0",
]
dev = [
    "black",
    "mink[test]",
    "mypy",
    "ruff",
]
all = [
    "mink[examples,dev]",
]

[project.urls]
Source = "https://github.com/kevinzakka/mink"
Tracker = "https://github.com/kevinzakka/mink/issues"
Changelog = "https://github.com/kevinzakka/mink/blob/main/CHANGELOG.md"
Homepage = "https://kevinzakka.github.io/mink/"
Documentation = "https://kevinzakka.github.io/mink/"

[tool.flit.module]
name = "mink"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle errors.
    "F",    # Pyflakes rules.
    "PLC",  # Pylint convention warnings.
    "PLE",  # Pylint errors.
    "PLR",  # Pylint refactor recommendations.
    "PLW",  # Pylint warnings.
    "I",    # Import sorting.
]
lint.ignore = [
    "E741",     # Ambiguous variable name. (l, O, or I)
    "E501",     # Line too long.
    "PLR2004",  # Magic value used in comparison.
    "PLR0915",  # Too many statements.
    "PLR0913",  # Too many arguments.
    "PLC0414",  # Import alias does not rename variable.
    "PLC1901",  # Use falsey strings.
    "PLR5501",  # Use `elif` instead of `else if`.
    "PLR0911",  # Too many return statements.
    "PLR0912",  # Too many branches.
    "PLW0603",  # Global statement updates are discouraged.
    "PLW2901",  # For loop variable overwritten.
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_configs = true
exclude = ["scripts/"]

[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "def __repr__",
    "if use_taylor:",
    "if theta_sq < get_epsilon(theta_sq\\.dtype):",
    "def ljac",
    "def rjac",
    "if mujoco.__version__"
]
omit = [
    "*exceptions.py",
    "*/contrib/*",
]
